# doc/OpenAPI/buttons/api/v1/auth.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "buttons/api/v1/auth.proto",
    "version": "1.0"
  },
  "tags": [
    {
      "name": "AuthService"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/auth/web/google_callback": {
      "get": {
        "summary": "get callback data",
        "description": "google send me callback data ",
        "operationId": "AuthService_GetWebGoogleCallback",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetWebGoogleCallbackResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "authuser",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "prompt",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "hd",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "OAauth 2.0"
        ]
      }
    },
    "/api/v1/auth/web/redirect_url": {
      "get": {
        "summary": "get redirect_url from provider",
        "description": "generate redirect_url and give to webview",
        "operationId": "AuthService_GetWebRedirectURL",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetWebRedirectURLResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "service",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "OAauth 2.0"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1GetWebGoogleCallbackResponse": {
      "type": "object"
    },
    "v1GetWebRedirectURLResponse": {
      "type": "object",
      "properties": {
        "provider": {
          "type": "string"
        },
        "redirectUrl": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "buttons-api gRPC-gateway repository",
    "url": "https://github.com/sundaytycoon/buttons-api"
  }
}